[REFACTORING] Split IMAPServerCompressWithSSLTest
